+++ templates/gpxcontent.php
@@ -282,6 +282,15 @@ echo '</p>'."\n";
 echo '<p id="gpxedit_version" style="display:none">';
 p($_['gpxedit_version']);
 echo '</p>'."\n";
+echo '<ul id="extrasymbols" style="display:none">';
+foreach($_['extrasymbols'] as $symbol){
+    echo '<li name="';
+    p($symbol['name']);
+    echo '">';
+    p($symbol['smallname']);
+    echo '</li>';
+}
+echo '</ul>'."\n";
 
 ?>
 </div>